+++ CHANGELOG.md
@@ -29,6 +29,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Record multiple parents per region [#27](https://github.com/Shopify/worldwide/pull/27)
 - Lookup by parent-child ISO and CLDR codes for dual-status territories
   [#29](https://github.com/Shopify/worldwide/pull/29)
+- Handle ISO_CODE only zones lookup [#26](https://github.com/Shopify/worldwide/pull/26)
 
 ---